+++ node/node.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"github.com/dgraph-io/badger/v3"
 	"github.com/gorilla/mux"
+	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/rovergulf/rbn/core"
 	"github.com/rovergulf/rbn/core/types"
 	"github.com/rovergulf/rbn/database/badgerdb"
@@ -15,7 +16,6 @@ import (
 	"github.com/spf13/viper"
 	"go.etcd.io/etcd/raft/v3"
 	"go.uber.org/zap"
-	"google.golang.org/grpc"
 	"os"
 	"sync"
 	"time"
@@ -44,8 +44,9 @@ type Node struct {
 	wm *wallets.Manager
 	db *badger.DB
 
-	grpcServer  *grpc.Server
+	host.Host
 	httpHandler httpServer
+	host        *host.Host
 
 	inGenRace bool
 
@@ -165,15 +166,16 @@ func (n *Node) Run(ctx context.Context) error {
 		"addr", nodeAddress, "is_root", n.metadata.Root)
 	go func() {
 		n.logger.Debugw("Listening gRPC", "addr", nodeAddress)
-		grpcSrv, err := n.PrepareGrpcServer()
+		h, err := n.PrepareP2pPeer(ctx)
 		if err != nil {
-			n.logger.Errorf("Unable to prepare gRPC server: %s", err)
+			n.logger.Errorf("Unable to prepare p2p host: %s", err)
 			n.Shutdown()
 		}
-		n.grpcServer = grpcSrv
 
-		if err := n.RunGrpcServer(nodeAddress); err != nil {
-			n.logger.Errorf("Unable to start gRPC server: %s", err)
+		n.Host = h
+
+		if err := n.RunP2pServer(ctx, ""); err != nil {
+			n.logger.Errorf("Unable to run p2p host: %s", err)
 			n.Shutdown()
 		}
 	}()
@@ -221,11 +223,13 @@ func (n *Node) Shutdown() {
 		}()
 	}
 
-	if n.grpcServer != nil {
+	if n.Host != nil {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			n.grpcServer.GracefulStop()
+			if err := n.Host.Close(); err != nil {
+				n.logger.Errorf("Unable to close p2p host: %s", err)
+			}
 		}()
 	}