+++ CHANGELOG.md
@@ -4,17 +4,24 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] v0.1.0
 
-## 14 Jul 2021
+## 4 Sep 2021
 
 ### Added
-- Reward transactions handle
+- libp2p usage
 
 ### Changed
+- pb files regenerated without gRPC server, probably should be removed
 
 ### Fixed
 
 ### Removed
+- grpc server usage
 
+
+## 14 Jul 2021
+
+### Added
+- Reward transactions handle
 
 ## 13 Jul 2021